Moved connector stories into text-input stories
